backend/requirements.txt
# Flask Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database
pymongo>=4.6.0
motor>=3.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# NLP & Embeddings
sentence-transformers>=2.2.0

# Kaggle Dataset Download
kagglehub>=0.1.0

# Environment
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0

# Utilities
tqdm>=4.65.0
